DjangoBackend/ecommerce_core/requirements.txt
pandas
openpyxl
Pillow
cloudinary
python-dotenv
requests
